chore: delete legacy clickByIndex + keyboard-simulation modules (#81)
Inline focusClickByIndex into keyboard.ts (the sole remaining consumer
of synthetic click for pre-keyboard focus). Delete src/lib/dom-actions/click.ts
and src/lib/keyboard-simulation.ts now that all consumers have been
migrated to CDP-based implementations. Remove clickByIndex export from
dom-actions/index.ts.
