+++ src/lib/agent/tools/keyboard.ts
@@ -18,7 +18,27 @@
 // Spec: docs/plans/2026-04-28-001-feat-phase2.5-cdp-keyboard-simulation-plan.md
 
 import type { ActionResult } from "../../dom-actions/types";
-import { clickByIndex } from "../../dom-actions/click";
+
+/**
+ * Self-contained function injected via chrome.scripting.executeScript to
+ * focus-click an element before a keyboard dispatch. Mirrors the old
+ * dom-actions/click.ts contract; lives here as the only remaining
+ * consumer of synthetic click (the public click tool is now CDP-based).
+ */
+function focusClickByIndex(index: number): ActionResult {
+  const el = document.querySelector(`[data-pie-idx="${index}"]`);
+  if (!el) {
+    return {
+      success: false,
+      error: `Element not found at index ${index}. The page may have changed; try snapshotting again.`,
+    };
+  }
+  (el as HTMLElement).click();
+  return {
+    success: true,
+    observation: `Focus-clicked element [${index}]`,
+  };
+}
 import { safeParseOrigin } from "../loop";
 import { requireCdpInput } from "./mouse";
 import type { CdpSession } from "../../../background/cdp-session";
@@ -257,7 +277,7 @@ export function buildKeyboardTools(deps: KeyboardToolDeps): Tool[] {
           const clickResult = await chrome.scripting
             .executeScript({
               target: { tabId: ctx.tabId },
-              func: clickByIndex,
+              func: focusClickByIndex,
               args: [a.after_element_index],
             })
             .then(
