+++ src/lib/dom-actions/index.ts
@@ -4,7 +4,6 @@ export type {
 } from "./types";
 
 // DOM action functions (injected into target page via executeScript)
-export { clickByIndex } from "./click";
 export { typeByIndex } from "./type";
 export { scroll } from "./scroll";
 export { selectByIndex } from "./select";
